Removed dashboard.original.
Some grep-ing around showed that only the unsaved changes
service used the original dashboard to check if the dashboard
changed, so I moved the responsibility to keep that thing around
into the unsaved changes service.

To do so, I gave dashboard-loaded an dashboard-saved the
loaded and saved dashboard as an event parameter, so I can
just copy the freshly loaded dashboard and keep it around
as the original. This also mostly decouples the unsavedChangesService
from the dashboard service, since it largely depends on the events.

I'll also need to check where I can use these events to do
more decoupling.