Made a major mess. Started to straighten out scopes & responsibilities.
I'm currently working out scopes and assignment of
dashboard/services/pulldowns/navs/editors/everything else and
controllers and in general I'm changing and modifying way too much
at once. But currently, this is like a giant ball of yarn, start pulling
on one end and the entire thing comes your way.

As a whole, I just want to load the dashboard into a scope where all
the parts of the dashboard management can access their specfic parts
of the dashboard, but currently, this is harder than it looks. I'll
keep digging though.

I'm just scared to merge all of this and I probably won't issue a
pull request until I have tested this internally for a week or two
while keeping up with upstream. Anything else would be way too
annoying to work with.